sysctlbyname( ) is expecting size as c_size_t
git-svn-id: https://svn.calendarserver.org/repository/calendarserver/CalendarServer/trunk@3346 e27351fd-9f3e-4f54-a53b-843176b1656c

+++ twistedcaldav/util.py
@@ -32,7 +32,7 @@ if sys.platform == "darwin" and hasCtypes:
 
     def getNCPU():
         ncpu = c_int(0)
-        size = c_int(sizeof(ncpu))
+        size = c_size_t(sizeof(ncpu))
 
         libc.sysctlbyname(
             "hw.ncpu",
